Reorganised the java_block_model code
Some of this had multiple uses so has been moved into the block package.
Others have been made part of the java resource pack manager class
